etl-pipeline-template/requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.26.0
requests>=2.31.0
pyyaml>=6.0
pyarrow>=14.0.0
apache-airflow>=2.7.0
pytest>=7.4.0
pytest-mock>=3.11.0
